test(lambda): merge account context tests into main lambda test file
